Implement rosout log widget.
Buffers formatted rosout messages in a ring buffer and displays them in
a box. Minimum loglevel, screen percentage and more are configurable.

In addition to enabling/disabling with 'L', I also added an intermediate
stopped state between enabled and disabled that can be useful to inspect
the last logs with a stopped buffer.

Known issues: log coloring sometimes doesn't work correctly, need to
investigate why (tui Text is weird). In the worst case you get white
text, so this is not a blocker atm.