Improve the `markers` module, support more shapes
Currently the module offers two listeners, one for markers array and one
for single marker topics. It only implements the ADD action and consider
all markers to be CUBEs.

We introduce here more supported shapes, support for the DELETE and
DELETEALL actions, as well as support for the marker's lifecycle. To
enable this, the markers from the array and single topics must be stored
in a common entity where they can be managed. To this end we introduce a
single listener `MarkersListener` which can subscribe to both topic
types. Both write the markers and action to a single new container.
Finally, the listener class has a timer which allows to delete markers
that expire.