Add the topic manager app mode
The Topic manager allows to add and remove topics from the config file.
Topics are sorted to the respective sub configs and only supported
topics are displayed.

The handling of changes is quite crude and requires a restart of termviz.
Also the sorting into the respective sub configs is hardcoded, to
prevent yet another breaking change in the config file.