Allow startup without active robot pose
Waits for a limited time (default 1s) for the robot pose to be
available, then asks the user if they want to start at the origin.

This is much more friendly than hanging and not starting up at all if the pose is not yet available, especially since you can initialize the pose via termviz.